Add files via upload
The very famous Snake game is designed using Matlab. It mainly invloved mix of algorithms and image processing.  The major task lies in defining the movements of snake and its consequences as it eats its food. The user is allowed to use "a" , "s" , "w" , "d" for left ,bottom, top and right respectively to move toward's the snake food. As soon as the food is eaten by the snake, the length of snake gets longer and the user has to make sure how he/she can get as many food items as possible without getting entangled between the head and tail of the snake.

Thank you! :)